+++ mocks_test.go
@@ -71,3 +71,5 @@ type mockBReader struct {
 func (r *mockBReader) Index() (IndexReader, error)          { return r.ir, nil }
 func (r *mockBReader) Chunks() (ChunkReader, error)         { return r.cr, nil }
 func (r *mockBReader) Tombstones() (TombstoneReader, error) { return newMemTombstones(), nil }
+func (r *mockBReader) MinTime() int64                       { return 0 } // Not implemented.
+func (r *mockBReader) MaxTime() int64                       { return 0 } // Not implemented.